services: Renamed "clock" -> "legacy" and "timestamp-clock" -> "clock"
