+++ addon/services/clock.js
@@ -1,51 +1,38 @@
 import Ember from 'ember';
-import TimestampClock from 'ember-cli-clock/services/timestamp-clock';
 
-export default TimestampClock.extend({
-  intervalTime: 1000,
-  second: 0,
-  minute: 0,
-  five: 0,
-  quarter: 0,
-  hour: 0,
+export default Ember.Service.extend({
+  interval: 1000,
 
-  reset() {
-    this.stop();
-    this.setProperties({second: 0, minute: 0, five: 0, quarter: 0, hour: 0});
+  time: null,
+  date: Ember.computed('time', function() {
+    return new Date(this.get('time'));
+  }),
+
+  init() {
+    this._super(...arguments);
     this.start();
   },
 
-  intervalChange: Ember.observer('intervalTime', function() {
-    if (Ember.testing) {
-      this.set('interval', this.get('intervalTime'));
-      return this.reset();
-    }
-    throw Error('The clock interval cannot be changed except during testing');
+  start() {
+    this.update();
+    this.set('intervalId', window.setInterval(() => this.update(), this.get('interval')));
+  },
+
+  stop() {
+    window.clearInterval(this.get('intervalId'));
+  },
+
+  willDestroy() {
+    this._super(...arguments);
+    this.stop();
+  },
+
+  onIntervalChange: Ember.observer('interval', function() {
+    this.stop();
+    this.start();
   }),
 
-  timeChange: Ember.observer('time', function() {
-    this.tick();
-  }),
-
-  tick() {
-    var second = this.incrementProperty('second');
-
-    if (second && (second % 60) === 0) {
-      var minute = this.incrementProperty('minute');
-
-      if (minute !== 0) {
-        if ((minute % 5) === 0) {
-          this.incrementProperty('five');
-        }
-
-        if ((minute % 15) === 0) {
-          this.incrementProperty('quarter');
-        }
-
-        if ((minute % 60) === 0) {
-          this.incrementProperty('hour');
-        }
-      }
-    }
+  update() {
+    Ember.run(() => this.set('time', Date.now()));
   }
 });
