+++ tests/dummy/app/services/clock-1sec.js
@@ -1,4 +1,4 @@
-import Clock from 'ember-cli-clock/services/timestamp-clock';
+import Clock from 'ember-cli-clock/services/clock';
 
 export default Clock.extend({
   interval: 1000
